ai-receptionist-pro/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.31
twilio==9.4.5
python-dateutil==2.9.0.post0
pytz==2024.1
httpx==0.27.0
